adding a new price curve
